Move ExecutionContextStore to a separate header
