+++ include/caffeine/Interpreter/Interpreter.h
@@ -14,6 +14,7 @@
 namespace caffeine {
 
 class ExecutionPolicy;
+class ExecutionContextStore;
 
 class ExecutionResult {
 public:
@@ -36,35 +37,6 @@ public:
 private:
   Status status_;
   llvm::SmallVector<Context, 2> contexts_;
-};
-
-/**
- * A store of Contexts that are not currently being executed.
- *
- * Note that when executing in a multithreaded fashion this may be accessed
- * concurrently from multiple threads.
- */
-class ExecutionContextStore {
-public:
-  ExecutionContextStore() = default;
-  virtual ~ExecutionContextStore() = default;
-
-  // Add a context to the store so that it can be retrieved later.
-  virtual void add_context(Context&& ctx) = 0;
-  // Get the next available context in the store
-  virtual Context next_context() = 0;
-
-  // Add multiple contexts to the store at once.
-  //
-  // By default this will just call add_context in a loop.
-  virtual void add_context_multi(Context* ctxs, size_t count);
-
-protected:
-  ExecutionContextStore(ExecutionContextStore&&) = default;
-  ExecutionContextStore(const ExecutionContextStore&) = default;
-
-  ExecutionContextStore& operator=(ExecutionContextStore&&) = default;
-  ExecutionContextStore& operator=(const ExecutionContextStore&) = default;
 };
 
 class Interpreter : public llvm::InstVisitor<Interpreter, ExecutionResult> {
@@ -115,6 +87,10 @@ public:
   ExecutionResult visitMemCpyInst(llvm::MemCpyInst& memcpy);
   ExecutionResult visitMemMoveInst(llvm::MemMoveInst& memmove);
   ExecutionResult visitMemSetInst(llvm::MemSetInst& memset);
+
+private:
+  void logFailure(const Context& ctx, const Assertion& assertion,
+                  std::string_view message = "");
 
 private:
   ExecutionResult visitExternFunc(llvm::CallInst& inst);
