+++ src/Interpreter/Interpreter.cpp
@@ -2,6 +2,7 @@
 #include "caffeine/Interpreter/ExprEval.h"
 #include "caffeine/Interpreter/Policy.h"
 #include "caffeine/Interpreter/StackFrame.h"
+#include "caffeine/Interpreter/Store.h"
 #include "caffeine/Interpreter/Value.h"
 #include "caffeine/Support/Assert.h"
 
@@ -20,12 +21,6 @@ namespace caffeine {
 ExecutionResult::ExecutionResult(Status status) : status_(status) {}
 ExecutionResult::ExecutionResult(llvm::SmallVector<Context, 2>&& contexts)
     : status_(Dead), contexts_(std::move(contexts)) {}
-
-void ExecutionContextStore::add_context_multi(Context* ctxs, size_t count) {
-  for (size_t i = 0; i < count; ++i) {
-    add_context(std::move(ctxs[i]));
-  }
-}
 
 Interpreter::Interpreter(Executor* queue, Context* ctx, FailureLogger* logger,
                          const InterpreterOptions& options)
